requirements.txt
streamlit
requests
beautifulsoup4
pandas
python-dotenv
python-telegram-bot
apscheduler